Add a way to update/delete note from the `gui/notes` tool
